book/rag_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
qdrant-client==1.7.0
python-dotenv==1.0.0
PyYAML==6.0.1
tiktoken==0.5.2
langchain==0.0.336
markdownify==0.11.6
aiofiles==23.2.1